--- a/.babelrc
+++ b/.babelrc
@@ -1,7 +1,11 @@
 {
   "presets": [
-    "es2015",
+    ["es2015", { "modules": false }],
     "react",
     "stage-0"
+  ],
+  "plugins": [
+    "transform-es2015-modules-commonjs",
+    "dynamic-import-node"
   ]
 }