--- a/actions/file_actions.jsx
+++ b/actions/file_actions.jsx
@@ -2,16 +2,13 @@
 // See License.txt for license information.
 
 import {batchActions} from 'redux-batched-actions';
-
 import request from 'superagent';
-
 import {FileTypes} from 'mattermost-redux/action_types';
 import {getLogErrorAction} from 'mattermost-redux/actions/errors';
 import {forceLogoutIfNecessary} from 'mattermost-redux/actions/helpers';
 import {Client4} from 'mattermost-redux/client';
 
 import store from 'stores/redux_store.jsx';
-
 import * as Utils from 'utils/utils.jsx';
 
 export function uploadFile(file, name, channelId, clientId, successCallback, errorCallback) {