--- a/actions/global_actions.jsx
+++ b/actions/global_actions.jsx
@@ -1,13 +1,12 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-import {browserHistory} from 'react-router';
-
 import {createDirectChannel, getChannelAndMyMember, getChannelStats, getMyChannelMember, joinChannel, viewChannel} from 'mattermost-redux/actions/channels';
 import {getPostThread} from 'mattermost-redux/actions/posts';
 import {removeUserFromTeam} from 'mattermost-redux/actions/teams';
 import {Client4} from 'mattermost-redux/client';
 
+import {browserHistory} from 'utils/browser_history';
 import {loadChannelsForCurrentUser} from 'actions/channel_actions.jsx';
 import {handleNewPost} from 'actions/post_actions.jsx';
 import {stopPeriodicStatusUpdates} from 'actions/status_actions.jsx';
@@ -21,14 +20,12 @@ import ErrorStore from 'stores/error_store.jsx';
 import store from 'stores/redux_store.jsx';
 import TeamStore from 'stores/team_store.jsx';
 import UserStore from 'stores/user_store.jsx';
-
 import WebSocketClient from 'client/web_websocket_client.jsx';
 
 import {ActionTypes, Constants, ErrorPageTypes, PostTypes} from 'utils/constants.jsx';
 import EventTypes from 'utils/event_types.jsx';
 import {sortTeamsByDisplayName} from 'utils/team_utils.jsx';
 import * as Utils from 'utils/utils.jsx';
-
 import en from 'i18n/en.json';
 import * as I18n from 'i18n/i18n.jsx';
 
@@ -104,8 +101,7 @@ export async function doFocusPost(channelId, postId, data) {
 
     dispatch({
         type: ActionTypes.RECEIVED_FOCUSED_POST,
-        data: postId,
-        channelId
+        data: postId
     });
 
     const member = getState().entities.channels.myMembers[channelId];
@@ -121,7 +117,7 @@ export function emitCloseRightHandSide() {
     dispatch(closeRightHandSide());
 }
 
-export async function emitPostFocusEvent(postId, onSuccess) {
+export async function emitPostFocusEvent(postId) {
     loadChannelsForCurrentUser();
     const {data} = await getPostThread(postId)(dispatch, getState);
 
@@ -134,11 +130,7 @@ export async function emitPostFocusEvent(postId, onSuccess) {
             loadNewGMIfNeeded(channel.id);
         }
 
-        doFocusPost(channelId, postId, data).then(() => {
-            if (onSuccess) {
-                onSuccess();
-            }
-        });
+        await doFocusPost(channelId, postId, data);
     } else {
         browserHistory.push('/error?type=' + ErrorPageTypes.PERMALINK_NOT_FOUND);
     }
@@ -460,7 +452,7 @@ export function clientLogout(redirectTo = '/') {
     stopPeriodicStatusUpdates();
     WebsocketActions.close();
     document.cookie = 'MMUSERID=;expires=Thu, 01 Jan 1970 00:00:01 GMT;';
-    window.location.href = redirectTo;
+    browserHistory.push(redirectTo);
 }
 
 export function toggleSideBarRightMenuAction() {
@@ -507,7 +499,7 @@ export async function redirectUserToDefaultTeam() {
         const channelId = BrowserStore.getGlobalItem(teamId);
         const channel = ChannelStore.getChannelById(channelId);
         if (channel) {
-            redirect(teams[teamId].name, channel);
+            redirect(teams[teamId].name, channel.name);
         } else if (channelId) {
             const {data} = await getChannelAndMyMember(channelId)(dispatch, getState);
             if (data) {