--- a/actions/integration_actions.jsx
+++ b/actions/integration_actions.jsx
@@ -2,7 +2,6 @@
 // See License.txt for license information.
 
 import request from 'superagent';
-
 import * as IntegrationActions from 'mattermost-redux/actions/integrations';
 import {getProfilesByIds} from 'mattermost-redux/actions/users';
 import {Client4} from 'mattermost-redux/client';
@@ -11,7 +10,6 @@ import AppDispatcher from 'dispatcher/app_dispatcher.jsx';
 import store from 'stores/redux_store.jsx';
 import TeamStore from 'stores/team_store.jsx';
 import UserStore from 'stores/user_store.jsx';
-
 import {ActionTypes} from 'utils/constants.jsx';
 import * as UserAgent from 'utils/user_agent.jsx';