--- a/actions/views/create_comment.jsx
+++ b/actions/views/create_comment.jsx
@@ -2,12 +2,10 @@
 // See License.txt for license information.
 
 import {createSelector} from 'reselect';
-
 import {getCurrentUserId} from 'mattermost-redux/selectors/entities/users';
 import {getCurrentTeamId} from 'mattermost-redux/selectors/entities/teams';
 import {makeGetMessageInHistoryItem, makeGetCommentCountForPost, getPost} from 'mattermost-redux/selectors/entities/posts';
 import {getCustomEmojisByName} from 'mattermost-redux/selectors/entities/emojis';
-
 import {
     addReaction,
     removeReaction,
@@ -15,7 +13,6 @@ import {
     moveHistoryIndexBack,
     moveHistoryIndexForward
 } from 'mattermost-redux/actions/posts';
-
 import {Posts} from 'mattermost-redux/constants';
 
 import * as PostActions from 'actions/post_actions.jsx';
@@ -23,9 +20,7 @@ import * as GlobalActions from 'actions/global_actions.jsx';
 import * as ChannelActions from 'actions/channel_actions.jsx';
 import {setGlobalItem, actionOnGlobalItemsWithPrefix} from 'actions/storage';
 import {EmojiMap} from 'stores/emoji_store.jsx';
-
 import {makeGetCommentDraft} from 'selectors/rhs';
-
 import * as Utils from 'utils/utils.jsx';
 import {Constants, StoragePrefixes} from 'utils/constants.jsx';