--- a/components/activity_log_modal/index.js
+++ b/components/activity_log_modal/index.js
@@ -3,7 +3,6 @@
 
 import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
-
 import {getSessions, revokeSession} from 'mattermost-redux/actions/users';
 
 import ActivityLogModal from './activity_log_modal.jsx';