--- a/components/admin_console/admin_navbar_dropdown/index.js
+++ b/components/admin_console/admin_navbar_dropdown/index.js
@@ -5,7 +5,6 @@ import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
 
 import {deferNavigation} from 'actions/admin_actions.jsx';
-
 import {getNavigationBlocked} from 'selectors/views/admin';
 
 import AdminNavbarDropdown from './admin_navbar_dropdown.jsx';