--- a/components/admin_console/admin_settings.jsx
+++ b/components/admin_console/admin_settings.jsx
@@ -5,9 +5,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 
 import {saveConfig} from 'actions/admin_actions.jsx';
-
 import {localizeMessage} from 'utils/utils.jsx';
-
 import SaveButton from 'components/save_button.jsx';
 import FormError from 'components/form_error.jsx';