--- a/components/admin_console/blockable_link/blockable_link.jsx
+++ b/components/admin_console/blockable_link/blockable_link.jsx
@@ -2,12 +2,12 @@
 // See License.txt for license information.
 
 import React from 'react';
-import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
+import {NavLink} from 'react-router-dom';
 
-import {Link, browserHistory} from 'react-router';
+import {browserHistory} from 'utils/browser_history';
 
-export default class BlockableLink extends React.PureComponent {
+export default class BlockableLink extends React.Component {
     static propTypes = {
 
         /*
@@ -33,10 +33,6 @@ export default class BlockableLink extends React.PureComponent {
         if (this.props.blocked) {
             e.preventDefault();
 
-            if (this.refs.link) {
-                ReactDOM.findDOMNode(this.refs.link).blur();
-            }
-
             this.props.actions.deferNavigation(() => {
                 browserHistory.push(this.props.to);
             });
@@ -49,10 +45,9 @@ export default class BlockableLink extends React.PureComponent {
         Reflect.deleteProperty(props, 'actions');
 
         return (
-            <Link
+            <NavLink
                 {...props}
                 onClick={this.handleClick}
-                ref='link'
             />
         );
     }