--- a/components/admin_console/blockable_link/index.js
+++ b/components/admin_console/blockable_link/index.js
@@ -5,7 +5,6 @@ import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
 
 import {deferNavigation} from 'actions/admin_actions.jsx';
-
 import {getNavigationBlocked} from 'selectors/views/admin';
 
 import BlockableLink from './blockable_link.jsx';