--- a/components/admin_console/cluster_settings.jsx
+++ b/components/admin_console/cluster_settings.jsx
@@ -3,7 +3,6 @@
 
 import React from 'react';
 import {FormattedHTMLMessage, FormattedMessage} from 'react-intl';
-
 import {Client4} from 'mattermost-redux/client';
 
 import * as Utils from 'utils/utils.jsx';