… .eslintrc.json
… components/admin_console/admin_console.jsx
… components/admin_console/custom_plugin_settings/custom_plugin_settings.jsx
… components/admin_console/custom_plugin_settings/index.js
… components/admin_console/data_retention_settings.jsx
… components/admin_console/database_settings.jsx
… components/admin_console/elasticsearch_settings.jsx
… components/admin_console/email_connection_test.jsx
… components/admin_console/email_settings.jsx
… components/admin_console/file_upload_setting.jsx
… components/admin_console/index.js
… components/admin_console/jobs/index.jsx
… components/admin_console/jobs/table.jsx
… components/admin_console/ldap_settings.jsx
… components/admin_console/license_settings.jsx
… components/admin_console/manage_roles_modal/index.js
… components/admin_console/manage_roles_modal/manage_roles_modal.jsx
… components/admin_console/manage_teams_modal/manage_teams_dropdown.jsx
… components/admin_console/manage_teams_modal/manage_teams_modal.jsx
… components/admin_console/manage_tokens_modal/index.js
… components/admin_console/manage_tokens_modal/manage_tokens_modal.jsx
… components/admin_console/plugin_management/index.js
… components/admin_console/plugin_management/plugin_management.jsx
… components/admin_console/remove_file_setting.jsx
… components/admin_console/reset_password_modal.jsx
… components/admin_console/revoke_token_button/index.js
… components/admin_console/saml_settings.jsx
… components/admin_console/server_logs/index.js
… components/admin_console/system_users/index.js
… components/admin_console/system_users/system_users.jsx