### Summary
- Updating react router to version 4
- Some rearrangement of webapp entry. Allows react to get to rendering faster.
- Routing style is different. Routes are now integrated into components rather than being seperate under routes.

### Developer impact
- Routes are now integrated into components. Check out [basic example](https://reacttraining.com/react-router/web/example/basic) to see how it's done now.
- The rest of the [react-router documentation](https://reacttraining.com/react-router/web/) is also helpfull if you need to deal with routing.
- Good example file for MM would be components/root.jsx
- browserHistory.push should be avoided. Use this.props.history.push on a router component when possible.
- React router does not parse the query parameters any more. Use [URLSearchParams ](https://developer.mozilla.org/en-US/docs/Web/API/URLSearchParams) for this.

Example:
```javascript
const clientId = this.props.location.query.client_id;
```
Becomes:
```javascript
const clientId = (new URLSearchParams(this.props.location.search)).get('client_id')
```

### Performance implications

Reported Load times | Before | After
-------------|--------|--------
Full speed | 500ms | 220ms
Fast 3G | 7.2s | 2.7s
6x CPU slowdown | 2.25s | 1.25s
6x CPU. Cache enabled | 2.7s | 1.5s


Stopwatch times | Before | After
-------------|--------|--------
Full speed | 2.0s | 1.3s
Fast 3G | 15.1s | 14.3s
6x CPU slowdown | 8.0s | 6.8s
6x CPU. Cache enabled | 7.7s | 6.6s

Sizes | Before | After
-------------|--------|--------
main bundle js | 1MB | 158KB
other bundles for channel view | 157.2KB | 810.1KB
total | 1157.2KB | 968.1KB

### Testing notes

This PR is very likely to cause bugs. Lots of the webapp code has been touched and the core of how the webapp navigates around has been changed. Testing should focus on:

- Redirects. Are you redirected where you expect?
- URL strangeness. Does the URL say what you expect?
- Backstage and the system console as they make heavy use of routing.
- Sign in and non-channel pages. These make heavy use of routing and are a pain for me to test.
- Bad links. Are you brought to where you expect when a bad link is clicked?