@mattermod 2018-01-18
Spinmint test server created at: http://i-013fcfc9edfe3c42a.spinmint.com

Test Account 1: Email: `test@test.com` | Password: `passwd`

Test Account 2: Email: `test2@test.com` | Password: `passwd`

Instance ID: i-013fcfc9edfe3c42a