@crspeller 2018-01-18
@jwilander @MusikPolice Feedback addressed.