@jasonblais 2018-01-18
Have reached out to @lindalumitchell for best way to test this. We might organize a small bug bash so we have more eyes on the changes at one time.