@jasonblais 2018-01-18
@crspeller Actually, a question on the performance implications (for changelog): what loadtimes are these? Page load time?