@jasonblais 2018-01-18
And what effects does the smaller main JS bundle have on performance (other than load times)?