@crspeller 2018-01-18
@jasonblais They are load times for already logged in and directly to town square.
The first set is measured by the browser and quite inaccurate for perception. But it's what the metrics we added measure.
The second set is a literal stopwatch until everything on the page is loaded.

A smaller main JS bundle allows our JavaScript code to start executing quicker. So we can do thing like display a loading screen if we want to quicker. Also it allows the browser to start doing some stuff in parallel.