@lindalumitchell 2018-01-18
JB had some testing notes:

1) Please pay close attention to the loading experience after clicking on different UI elements.
It seems the first load of a few modals (Account Settings, More Direct Messages menu) appear broken: the text of the modal loads before the background, which loads a split second later.

And a question for you in JB's second note, @crspeller:

2) Chris mentioned that we load the Katex library only when the user needs it (the library is for supporting Latex formatting). However, LB noticed that Latex didn't work for her on the server. So it might be useful to ask Chris what other libraries are no longer loaded during the initial load and test those too.