@lindalumitchell 2018-01-18
Also, @crspeller, a couple issues so far (let me know if you want a ticket or if commenting like this is good):

1. Screen flashes on some pages (page starts to load, then looks like it refreshes and loads again). So far seeing it consistently on logout and on finishing the tutorial. Screen recording of finishing the tutorial:

![pr-602-screen-flash](https://user-images.githubusercontent.com/25826533/35126861-2e2eac66-fc64-11e7-99d5-1f0ab3091e9a.gif)

2. Can't create a new team (two issues):
 a. Main menu > Create a New Team
 b. Enter team name and press Enter or click Next
 c. Observed: URL field is not pre-populated. Expected: URL field is pre-populated based on name entered in step b.
 d. Enter a valid team URL (lowercase, numbers, dashes) and press Enter or click Finish
 e. Observed: Error "Invalid name", 400 error in console and in log, can't continue. Expected: If valid entry in the URL field, new team is created and user is redirected to new team Town Square

![errors-on-finishing-team-create](https://user-images.githubusercontent.com/25826533/35126940-8fb17c7a-fc64-11e7-9629-5c5e0129bc07.png)

3. System Console LHS scrolls itself up after user clicks LHS link
 a. Open System Console
 b. Scroll the LHS down and click any link toward the bottom of the LHS
 c. Observed: Page opens, LHS scrolls back up to the top, message in console about "scroll-blocking 'wheel' event". Expected: Page opens, LHS stays where it is

![sys-console-lhs-scrolls-up-on-click](https://user-images.githubusercontent.com/25826533/35127260-287c7f9e-fc66-11e7-8a58-02bc20d81b4e.gif)