@crspeller 2018-01-19
@lindalumitchell I think I have fixed everything listed. With the exception of "It seems the first load of a few modals appear broken" because I couldn't reproduce any problem. Models load the same on master as they do here.