@crspeller 2018-01-19
Was just testing and my fixes seem to have broken permalinks.