@mattermod 2018-01-19
Spinmint test server created at: http://i-0fb23848d0c5e9495.spinmint.com

Test Account 1: Email: `test@test.com` | Password: `passwd`

Test Account 2: Email: `test2@test.com` | Password: `passwd`

Instance ID: i-0fb23848d0c5e9495