@lindalumitchell 2018-01-19
Finished testing many functions around redirects (plus a bunch of basic functionality). Will need to do detailed testing of permalinks after the issues CS and I noticed earlier are fixed. A list of issues CS and I have chatted about (struck through if already fixed):

([testing details in spreadsheet](https://docs.google.com/spreadsheets/d/1-fkoCtxXUahqkEc6geMCqXjlsWdC_IC7vQxHEPoxyXQ/edit?usp=sharing)):

1. Stutter on page load (noticed on logout ~~and finishing tutorial~~)
2. ~~Team creation: URL field isn't pre-populated, and can't save new team~~
3. ~~System Console scrolls itself up after LHS click~~
4. System Console radio buttons require two clicks and pasting text requires two CMD+Vs
5. Permalink: Pasting a permalink in browser redirects to Town Square (actually, pasting any link in browser or even simply refreshing does this)
6. Permalink: Jump from search results does not open permalink as expected (see spreadsheet for more detail on different behaviors)
7. Clicking desktop notification redirects to Town Square, not to the mention or DM
8. Refresh causes error in console (`http://i-0fb23848d0c5e9495.spinmint.com/api/v4/roles/names 404 (Not Found)`) (maybe related to it always redirecting to Town Square?), and sometimes shows blank page with additional errors:
  - TypeError: Cannot read property 'top' of undefined, Uncaught (in promise) 
    -- at t.J.updateUnreadIndicators
    -- at t.value
    -- at updateClassInstance
    -- at beginWork
  - TypeError: Cannot read property 'top' of undefined)
    -- (same as previous)