@mattermod 2018-01-19
Spinmint test server destroyed