@mattermod 2018-01-19
Spinmint test server created at: http://i-084d29e754765619b.spinmint.com

Test Account 1: Email: `test@test.com` | Password: `passwd`

Test Account 2: Email: `test2@test.com` | Password: `passwd`

Instance ID: i-084d29e754765619b