@lindalumitchell 2018-01-20
Updated issues list (have shared screenshots with CS in a DM but can add them here as well if helpful, just let me know):

1. ~~Stutter on page load (noticed on logout and finishing tutorial~~)
2. ~~Team creation: URL field isn't pre-populated, and can't save new team~~
3. ~~System Console scrolls itself up after LHS click~~
4. System Console radio buttons require two clicks and pasting text requires two CMD+Vs
5. ~~Permalink: Pasting a permalink in browser redirects to Town Square (actually, pasting any link in browser or even simply refreshing does this)~~
6. ~~Permalink: Jump from search results does not open permalink as expected (see spreadsheet for more detail on different behaviors)~~
7. Clicking desktop notification ~~redirects to Town Square, not to the mention or DM~~ causes endless looping refreshes of Town Square and blank page, then that user sees a blank page on trying to log in again
8. ~~Sometimes refresh shows blank page with errors in console (screenshot errors-on-refresh.png attached in message below)~~
9. Still seeing 404s in console on refresh, even if page loads correctly
10. Permalink to deleted message (or message in a private channel the user isn't a member of) tries to show error page but the page is blank
11. Trying to open URL of a private channel the user isn't a member of causes endless looping refreshes of Town Square and blank page
12. Log out, open permalink URL, enter credentials, opens to Town Square (should open to permalink)