@mattermod 2018-01-21
Spinmint test server destroyed